+++ nextgisweb/pyramid/nodepkg/metrics/index.ts
@@ -1,0 +1,3 @@
+/** @entrypoint */
+import { MetricsSettings } from "./MetricSettings";
+export default MetricsSettings;