+++ nextgisweb/pyramid/view.py
@@ -198,6 +198,13 @@ def home_path(request):
         title=_("Home path"),
         dynmenu=request.env.pyramid.control_panel)
 
+@viewargs(renderer='react')
+def metrics(request):
+    request.require_administrator()
+    return dict(
+        entrypoint='@nextgisweb/pyramid/metrics',
+        title=_("Metrics and analytics"),
+        dynmenu=request.env.pyramid.control_panel)
 
 def test_request(request):
     comp = request.env.pyramid
@@ -460,6 +467,11 @@ def setup_pyramid(comp, config):
         '/control-panel/home-path'
     ).add_view(home_path)
 
+    config.add_route(
+        'pyramid.control_panel.metrics',
+        '/control-panel/metrics',
+    ).add_view(metrics)
+
     config.add_route('pyramid.locale', '/locale/{locale:str}').add_view(locale)
 
     comp.test_request_handler = None
@@ -491,6 +503,8 @@ def setup_pyramid(comp, config):
             args.request.route_url('pyramid.control_panel.logo'))),
         dm.Link('settings/home_path', _("Home path"), lambda args: (
             args.request.route_url('pyramid.control_panel.home_path'))),
+        dm.Link('settings/metrics', _("Metrics and analytics"), lambda args: (
+            args.request.route_url('pyramid.control_panel.metrics'))),
     )
 
     if env.core.options['storage.enabled']: