+++ CHANGES.rst
@@ -15,6 +15,7 @@ Changes
 - Ant Design library upgraded to 5.
 - Save to PDF from the webmap printing panel.
 - Store audit journal in PostgreSQL database instead of ElasticSearch.
+- Support for Google Analytics and Yandex.Metrica metrics.
 
 
 4.4.0